Added shakespeare LSTM
