+++ fedml_api/model/Bsaeline/CIFAR10.py
@@ -1,0 +1,74 @@
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+import torch.optim as optim
+from torchvision import datasets, transforms
+from torch.optim.lr_scheduler import StepLR
+import copy
+
+
+
+class Cifar10_Net(nn.Module):
+    def __init__(self):
+        super(Cifar10_Net, self).__init__()
+        # convolutional layer (sees 32x32x3 image tensor)
+        self.conv1 = nn.Conv2d(3, 16, 3, padding=1)
+        # convolutional layer (sees 16x16x16 tensor)
+        self.conv2 = nn.Conv2d(16, 32, 3, padding=1)
+        # convolutional layer (sees 8x8x32 tensor)
+        self.conv3 = nn.Conv2d(32, 64, 3, padding=1)
+        # max pooling layer
+        self.pool = nn.MaxPool2d(2, 2)
+        # linear layer (64 * 4 * 4 -> 500)
+        self.fc1 = nn.Linear(64 * 4 * 4, 500)
+        # linear layer (500 -> 10)
+        self.fc2 = nn.Linear(500, 10)
+        # dropout layer (p=0.25)
+        self.dropout = nn.Dropout(0.25)
+
+    def forward(self, x):
+        # add sequence of convolutional and max pooling layers
+        x = self.pool(F.relu(self.conv1(x)))
+        x = self.pool(F.relu(self.conv2(x)))
+        x = self.pool(F.relu(self.conv3(x)))
+        # flatten image input
+        x = x.view(-1, 64 * 4 * 4)
+        # add dropout layer
+        x = self.dropout(x)
+        # add 1st hidden layer, with relu activation function
+        x = F.relu(self.fc1(x))
+        # add dropout layer
+        x = self.dropout(x)
+        # add 2nd hidden layer, with relu activation function
+        x = self.fc2(x)
+        return x
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
