+++ test.py
@@ -1,9 +1,8 @@
 '''
-from fedml_api.data_preprocessing.shakespeare.data_loader import load_partition_data_shakespeare
+from fedml_api.data_preprocessing.load_data import load_partition_data_shakespeare
 
 
 (
-    client_num,
     train_data_num,
     test_data_num,
     train_data_global,
@@ -12,20 +11,9 @@ from fedml_api.data_preprocessing.shakespeare.data_loader import load_partition_
     train_data_local_dict,
     test_data_local_dict,
     output_dim,
-) = load_partition_data_shakespeare(10)
+) = load_partition_data_shakespeare(50,"data/shakespeare")
 
 
-
-print("=============================client_num")
-print(client_num)
-print("=========================train_data_num")
-print(train_data_num)
-print("==========================test_data_num")
-print(test_data_num)
-print("=============================output_dim")
-print(output_dim)
-print("loaded")
-'''
